requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.32.4 